+++ src/components/ConditionalFilter/FiltersQueryBuilder/queryVarsBuilders/BaseMappableQueryVarsBuilder.ts
