+++ src/components/ConditionalFilter/FiltersQueryBuilder/queryVarsBuilders/StaticBooleanQueryVarsBuilder.test.ts
