+++ src/components/ConditionalFilter/FiltersQueryBuilder/queryVarsBuilders/StaticBooleanQueryVarsBuilder.ts
