+++ src/components/ConditionalFilter/FiltersQueryBuilder/queryVarsBuilders/StaticQueryVarsBuilder.test.ts
