+++ src/components/ConditionalFilter/FiltersQueryBuilder/queryVarsBuilders/StaticQueryVarsBuilder.ts
